ui fix

+++ core/src/main/java/me/carina/rpg/common/stat/Multiplier.java
@@ -1,8 +1,8 @@
 package me.carina.rpg.common.stat;
 
 public class Multiplier extends StatusEffect{
-    float addAmount = 0;
-    float mulAmount = 0;
+    int addAmount = 0;
+    int mulAmount = 0;
     Affinity targetAffinity;
     StatType targetStat;
 }
